+++ cache/remotecache/export.go
@@ -1,4 +1,4 @@
-package cacheimport
+package remotecache
 
 import (
 	"bytes"
@@ -10,7 +10,7 @@ import (
 	"github.com/containerd/containerd/content"
 	"github.com/containerd/containerd/images"
 	"github.com/docker/distribution/manifest"
-	v1 "github.com/moby/buildkit/cache/cacheimport/v1"
+	v1 "github.com/moby/buildkit/cache/remotecache/v1"
 	"github.com/moby/buildkit/session"
 	solver "github.com/moby/buildkit/solver-next"
 	"github.com/moby/buildkit/util/contentutil"
