+++ control/control.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/docker/distribution/reference"
 	controlapi "github.com/moby/buildkit/api/services/control"
-	"github.com/moby/buildkit/cache/cacheimport"
+	"github.com/moby/buildkit/cache/remotecache"
 	"github.com/moby/buildkit/client"
 	"github.com/moby/buildkit/exporter"
 	"github.com/moby/buildkit/frontend"
@@ -25,8 +25,8 @@ type Opt struct {
 	WorkerController *worker.Controller
 	Frontends        map[string]frontend.Frontend
 	CacheKeyStorage  solver.CacheKeyStorage
-	CacheExporter    *cacheimport.CacheExporter
-	CacheImporter    *cacheimport.CacheImporter
+	CacheExporter    *remotecache.CacheExporter
+	CacheImporter    *remotecache.CacheImporter
 }
 
 type Controller struct { // TODO: ControlService
@@ -149,7 +149,7 @@ func (c *Controller) Solve(ctx context.Context, req *controlapi.SolveRequest) (*
 		}
 	}
 
-	var cacheExporter *cacheimport.RegistryCacheExporter
+	var cacheExporter *remotecache.RegistryCacheExporter
 	if ref := req.Cache.ExportRef; ref != "" {
 		parsed, err := reference.ParseNormalizedNamed(ref)
 		if err != nil {
